Added tracy as dependency + some instrumentation
To use tracy add TRACY_ENABLE and RelWithDebInfo as cmake args, run the
tests with sudo and connect with the tracy client.
Note: Collecting to many trace info while not connected with the client
will eventually cause the app to be killed, so don't forget to connect
if you enabled tracy.